wt88047: Fix SIGFPE when closing camera or taking a picture
There's a division by zero when tearing down the camera. The origin of
the divisor value seems to be an uninitialized variable, so it probably
worked out of luck only in cm-12.1.

Adapted for WT88047 by Ketut Putu Kumajaya <ketut.kumajaya@gmail.com>

The original patch for R7 plus, not all changes applied to WT88047 but
this really a fix for SIGFPE

Change-Id: Ieb86cab882d35b2889a056929ca43cc15f303f08

Signed-off-by: RazorSharp <kirananto@gmail.com>